requirements.txt
pyqt5==5.13.1
google-auth==2.30.0
google-api-python-client==2.135.0